+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 4.6.0 (2024-??-??)
 
 - Add support for `ruflin/elastica` 8.x
+- Deprecate not returning a query builder from the modifier callback in `Pagerfanta\Doctrine\DBAL\QueryAdapter`
 
 ## 4.5.0 (2024-04-10)